Move `compiler::Target` to `core`
