+++ lib/vrl/compiler/src/lib.rs
@@ -1,7 +1,6 @@
 mod compiler;
 mod context;
 mod program;
-mod target;
 mod test_util;
 
 pub mod expression;
@@ -13,13 +12,13 @@ pub mod vm;
 
 pub use crate::value::Value;
 pub use context::Context;
+pub use core::{value, Target};
 pub(crate) use diagnostic::Span;
 pub use expression::{Expression, ExpressionError, Resolved};
 pub use function::{Function, Parameter};
 pub use paste::paste;
 pub use program::Program;
 pub(crate) use state::Compiler as State;
-pub use target::Target;
 pub use type_def::TypeDef;
 
 pub type Result = std::result::Result<Program, compiler::Errors>;
