+++ lib/vrl/compiler/src/value.rs
@@ -2,8 +2,6 @@ mod arithmetic;
 mod convert;
 mod error;
 pub mod kind;
-mod r#macro;
-mod target;
 
 pub use error::Error;
 pub use kind::{Collection, Field, Index, Kind};
