+++ lib/vrl/core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "vrl-core"
+version = "0.1.0"
+authors = ["Vector Contributors <vector@datadoghq.com>"]
+edition = "2021"
+publish = false
+
+[dependencies]
+lookup = { path = "../../lookup" }
+value = { path = "../../value" }
