+++ lib/vrl/core/src/lib.rs
@@ -1,0 +1,5 @@
+mod r#macro;
+mod target;
+
+pub use target::Target;
+pub use value::Value;
