+++ lib/vrl/vrl/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 
 [dependencies]
 bytes = "1.1.0"
+core = { package = "vrl-core", path = "../core" }
 compiler = { package = "vrl-compiler", path = "../compiler" }
 diagnostic = { package = "vrl-diagnostic", path = "../diagnostic" }
 indoc = "1"
